refactor(ts): service helper functions
